In preparation for writing unit tests, and to make the source sane again... I split the subsystems into separate source files, local modules for curve1174
