+++ crypto/src/curve1174/mod.rs
@@ -36,2089 +36,29 @@ use std::ops::{Add, Sub, Mul, Div, Neg, AddAssign, SubAssign, MulAssign, DivAssi
 use hash::*;
 use std::cmp::Ordering;
 
-// -----------------------------------------------------------------------------------
+mod winvec;
+use self::winvec::*;
 
-const WINDOW : usize  =  4; // using 4-bit fixed windows
+mod lev32;
+use self::lev32::*;
 
-const BOT_51_BITS : i64 = ((1 << 51) - 1); // Fq51 frames contain 51 bits
-const BOT_47_BITS : i64 = ((1 << 47) - 1); // MSB frame only has 47 bits
-const CURVE_D     : i64 = -1174; // the d value in the curve equation
-const CURVE_H     : i64 = 4; // cofactor of curve group
+mod u256;
+use self::u256::*;
 
-// nbr of precomputed point multiples
-const NPREP: usize = 1 + (1 << (WINDOW-1));
+mod fr;
+use self::fr::*;
 
-// -----------------------------------------------------------------
-// window vector of 4-bit values, used for fast multiply of curve points
+mod fq;
+use self::fq::*;
 
-const PANES  : usize  = 64; // nbr of 4-bit nibbles in 256-bit numbers
+mod fq51;
+use self::fq51::*;
 
-#[repr(C)]
-struct WinVec([i8;PANES]);
+mod ecpt;
+use self::ecpt::*;
 
-const WINVEC_INIT: WinVec = WinVec([0;PANES]);
-
-impl From<FrUnscaled> for WinVec {
-    fn from(x : FrUnscaled) -> WinVec {
-        let tmp = LEV32::from(x.0);
-        let mut wv = WINVEC_INIT;
-        cwin4(&tmp, &mut wv);
-        wv
-    }
-}
-
-impl From<i64> for WinVec {
-    fn from(x : i64) -> WinVec {
-        WinVec::from(FrUnscaled::from(x))
-    }
-}
-
-impl From<Fr> for WinVec {
-    fn from(x : Fr) -> WinVec {
-        WinVec::from(FrUnscaled::from(x))
-    }
-}
-
-// -----------------------------------------------------------------
-// type LEV32 represents a 256-bit bignum as a little-endian 32-byte vector
-
-#[derive(Copy, Clone)]
-#[repr(C)]
-struct LEV32([u8;32]);
-
-impl fmt::Display for LEV32 {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "LEV32({})", self.nbr_str())
-    }
-}
-
-impl LEV32 {
-    fn nbr_str(&self) -> String {
-        let LEV32(qv) = self;
-        let v = unsafe { mem::transmute::<[u8;32], [u64;4]>(*qv) };
-        basic_nbr_str(&v)
-    }
-}
-
-// -----------------------------------------------------------------
-// U256 word chunks represent a 256-bit bignum as a little-endian u64 vector
-
-#[derive(Copy, Clone, Eq, PartialEq)]
-#[repr(C)]
-struct U256([u64;4]);
-
-impl From<LEV32> for U256 {
-    fn from(v : LEV32) -> U256 {
-        let xv = unsafe { mem::transmute::<[u8;32], [u64;4]>(v.0) };
-        U256(xv)
-    }
-}
-
-impl From<U256> for LEV32 {
-    fn from(v : U256) -> LEV32 {
-        let xv = unsafe { mem::transmute::<[u64;4], [u8;32]>(v.0) };
-        LEV32(xv)
-    }
-}
-
-impl U256 {
-    fn zero() -> U256 {
-        U256([0;4])
-    }
-
-    fn one() -> U256 {
-        U256([1,0,0,0])
-    }
-
-    fn nbr_str(&self) -> String {
-        basic_nbr_str(&self.0)
-    }
-
-    fn add_mod(&mut self, other: &U256, modulo: &U256) {
-        add_nocarry(&mut self.0, &other.0);
-        if *self >= *modulo {
-            sub_noborrow(&mut self.0, &modulo.0);
-        }
-    }
-
-    fn sub_mod(&mut self, other: &U256, modulo: &U256) {
-        if *self < *other {
-            add_nocarry(&mut self.0, &modulo.0);
-        }
-        sub_noborrow(&mut self.0, &other.0);
-    }
-
-    fn neg_mod(&mut self, modulo: &U256) {
-        if *self > Self::zero() {
-            let mut tmp = modulo.0;
-            sub_noborrow(&mut tmp, &self.0);
-            self.0 = tmp;
-        }
-    }
-
-    /// Multiply `self` by `other` (mod `modulo`) via the Montgomery
-    /// multiplication method.
-    fn mul_mod(&mut self, other: &U256, modulo: &U256, inv: u64) {
-        mul_reduce(&mut self.0, &other.0, &modulo.0, inv);
-
-        if *self >= *modulo {
-            sub_noborrow(&mut self.0, &modulo.0);
-        }
-    }
-
-    #[inline]
-    fn is_even(&self) -> bool {
-        self.0[0] & 1 == 0
-    }
-
-    /// Turn `self` into its multiplicative inverse (mod `modulo`)
-    fn invert_mod(&mut self, modulo: &U256) {
-        // Guajardo Kumar Paar Pelzl
-        // Efficient Software-Implementation of Finite Fields with Applications to Cryptography
-        // Algorithm 16 (BEA for Inversion in Fp)
-
-        let mut u = *self;
-        let mut v = *modulo;
-        let mut b = U256::one();
-        let mut c = U256::zero();
-
-        while u != U256::one() && v != U256::one() {
-            while u.is_even() {
-                div2(&mut u.0);
-
-                if b.is_even() {
-                    div2(&mut b.0);
-                } else {
-                    add_nocarry(&mut b.0, &modulo.0);
-                    div2(&mut b.0);
-                }
-            }
-            while v.is_even() {
-                div2(&mut v.0);
-
-                if c.is_even() {
-                    div2(&mut c.0);
-                } else {
-                    add_nocarry(&mut c.0, &modulo.0);
-                    div2(&mut c.0);
-                }
-            }
-
-            if u >= v {
-                sub_noborrow(&mut u.0, &v.0);
-                b.sub_mod(&c, modulo);
-            } else {
-                sub_noborrow(&mut v.0, &u.0);
-                c.sub_mod(&b, modulo);
-            }
-        }
-
-        if u == U256::one() {
-            self.0 = b.0;
-        } else {
-            self.0 = c.0;
-        }
-    }
-}
-
-impl Ord for U256 {
-    fn cmp(&self, other: &U256) -> Ordering {
-        for (a, b) in self.0.iter().zip(other.0.iter()).rev() {
-            if *a < *b {
-                return Ordering::Less;
-            }
-            else if *a > *b {
-                return Ordering::Greater;
-            }
-        }
-        return Ordering::Equal;
-    }
-}
-
-impl PartialOrd for U256 {
-    fn partial_cmp(&self, other: &U256) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl fmt::Debug for U256 {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "U256([{:016x}, {:016x}, {:016x}, {:016x}])", 
-            self.0[0] as u64,
-            self.0[1] as u64,
-            self.0[2] as u64,
-            self.0[3] as u64)
-    }
-}
-
-impl fmt::Display for U256 {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "U256({})", self.nbr_str())
-    }
-}
-
-impl From<FrUnscaled> for U256 {
-    fn from(x: FrUnscaled) -> U256 {
-        x.0
-    }
-}
-
-impl From<Fr> for FrUnscaled {
-    fn from(x: Fr) -> FrUnscaled {
-        let mut tmp = x.0;
-        mul_collapse(&mut tmp.0, &R.0, FRINV);
-        FrUnscaled(tmp)
-    }
-}
-
-impl From<FqUnscaled> for U256 {
-    fn from(x: FqUnscaled) -> U256 {
-        x.0
-    }
-}
-
-impl From<Fq> for FqUnscaled {
-    fn from(x: Fq) -> FqUnscaled {
-        let mut tmp = x.0;
-        mul_collapse(&mut tmp.0, &Q.0, FQINV);
-        FqUnscaled(tmp)
-    }
-}
-
-// -------------------------------------------------------------------------
-// Primitive ops for U256, on raw arrays of 4 x u64
-
-/// Divide by two
-#[inline]
-fn div2(a: &mut [u64; 4]) {
-    let mut t = a[3] << 63;
-    a[3] = a[3] >> 1;
-    let b = a[2] << 63;
-    a[2] >>= 1;
-    a[2] |= t;
-    t = a[1] << 63;
-    a[1] >>= 1;
-    a[1] |= b;
-    a[0] >>= 1;
-    a[0] |= t;
-}
-
-/// Multiply by two
-#[inline]
-fn mul2(a: &mut [u64; 4]) {
-    let mut last = 0;
-    for i in a {
-        let tmp = *i >> 63;
-        *i <<= 1;
-        *i |= last;
-        last = tmp;
-    }
-}
-
-fn split_u64(i : u64) -> (u64, u64) {
-    (i >> 32, i & ((1 << 32) - 1))
-}
-
-fn combine_u64(hi: u64, lo: u64) -> u64 {
-    (hi << 32) | lo
-}
-
-fn adc(a: u64, b: u64, carry: &mut u64) -> u64 {
-    let (a1, a0) = split_u64(a);
-    let (b1, b0) = split_u64(b);
-    let (c, r0) = split_u64(a0 + b0 + *carry);
-    let (c, r1) = split_u64(a1 + b1 + c);
-    *carry = c;
-
-    combine_u64(r1, r0)
-}
-
-fn add_nocarry(a: &mut [u64;4], b: &[u64; 4]) {
-    let mut carry = 0;
-    for (a, b) in a.into_iter().zip(b.iter()) {
-        *a = adc(*a, *b, &mut carry);
-    }
-    debug_assert!(0 == carry);
-}
-
-fn sub_noborrow(a: &mut [u64;4], b: &[u64;4]) {
-    fn sbb(a: u64, b: u64, borrow: &mut u64) -> u64 {
-        let (a1,a0) = split_u64(a);
-        let (b1,b0) = split_u64(b);
-        let (b, r0) = split_u64((1 << 32) + a0 - b0 - *borrow);
-        let (b, r1) = split_u64((1 << 32) + a1 - b1 - ((b == 0) as u64));
-        *borrow = (b == 0) as u64;
-        combine_u64(r1, r0)
-    }
-
-    let mut borrow = 0;
-    for (a, b) in a.into_iter().zip(b.iter()) {
-        *a = sbb(*a, *b, &mut borrow);
-    }
-    debug_assert!(0 == borrow);
-}
-
-fn mac_digit(acc: &mut [u64], b: &[u64], c: u64)
-{
-    #[inline]
-    fn mac_with_carry(a: u64, b: u64, c: u64, carry: &mut u64) -> u64 {
-        let (b_hi, b_lo) = split_u64(b);
-        let (c_hi, c_lo) = split_u64(c);
-
-        let (a_hi, a_lo) = split_u64(a);
-        let (carry_hi, carry_lo) = split_u64(*carry);
-        let (x_hi, x_lo) = split_u64(b_lo * c_lo + a_lo + carry_lo);
-        let (y_hi, y_lo) = split_u64(b_lo * c_hi);
-        let (z_hi, z_lo) = split_u64(b_hi * c_lo);
-        let (r_hi, r_lo) = split_u64(x_hi + y_lo + z_lo + a_hi + carry_hi);
-
-        *carry = (b_hi * c_hi) + r_hi + y_hi + z_hi;
-
-        combine_u64(r_lo, x_lo)
-    }
-
-    if c == 0 {
-        return;
-    }
-
-    let mut b_iter = b.iter();
-    let mut carry = 0;
-
-    for ai in acc.iter_mut() {
-        if let Some(bi) = b_iter.next() {
-            *ai = mac_with_carry(*ai, *bi, c, &mut carry);
-        } else if carry != 0 {
-            *ai = mac_with_carry(*ai, 0, c, &mut carry);
-        } else {
-            break;
-        }
-    }
-
-    debug_assert!(carry == 0);
-}
-
-#[inline]
-fn mul_reduce(
-    this: &mut [u64; 4],
-    by: &[u64; 4],
-    modulus: &[u64; 4],
-    inv: u64
-)
-{
-    // This function produces the Montgomery reduction of the product
-    // of a = `this` and b = `by`, modulo m = `modulus`. That means that 
-    // we get
-    //
-    //       result = ((a * b)/Z) mod m
-    //
-    // where Z = 2^256 mod m. Parameter `inv` = -1/m mod 2^64.
-    //
-    // If a and b are pre-scaled by Z, then so is their product.
-    // To convert back to raw form (non-prescale), perform a multiply with
-    // a unit vector [1,0,0,0].
-    //
-    // The Montgomery reduction here is based on Algorithm 14.32 in
-    // Handbook of Applied Cryptography
-    // <http://cacr.uwaterloo.ca/hac/about/chap14.pdf>.
-
-    let mut res = [0; 2*4];
-    for (i, xi) in this.iter().enumerate() {
-        // this produces a double-wide product in `res`
-        mac_digit(&mut res[i..], by, *xi);
-    }
-
-    for i in 0..4 {
-        // Montgomery reduction
-        let k = inv.wrapping_mul(res[i]);
-        mac_digit(&mut res[i..], modulus, k);
-    }
-
-    this.copy_from_slice(&res[4..]);
-}
-
-fn mul_collapse(this: &mut [u64;4], modulus: &[u64;4], inv: u64) {
-    // this is a short circuiting function for removing the Montgomery prescaling
-    // slightly faster than a full multiply by [1,0,0,0]
-    let mut res = [0; 2*4];
-    for (i, xi) in this.iter().enumerate() {
-        res[i] = *xi;
-    }
-    for i in 0..4 {
-        // Montgomery reduction
-        let k = inv.wrapping_mul(res[i]);
-        mac_digit(&mut res[i..], modulus, k);
-    }
-    this.copy_from_slice(&res[4..]);
-}
-
-// produce a 256-bit bignum string from a LEV of [u64;4]
-fn basic_nbr_str(x: &[u64;4]) -> String {
-    format!("{:016x}{:016x}{:016x}{:016x}", 
-        x[3], x[2], x[1], x[0])
-}
-
-// -----------------------------------------------------------------
-// Fr is the field on the curve.  |Fr| * GenPt = INF
-// |Fr| < |Fq|, both |Fr| and |Fq| are prime.
-//
-// Type Fr is for working directly in the field Fr, using fast Montgomery reduction
-// for modular multiply. As such, Fr is scaled by the R_ONE value. 
-//
-// Point multiplication on the curve is performed using windowed vectors of 
-// unscaled Fr values. For that we have type FrUnscaled, to avoid the overhead
-// of scaling / descaling.
-
-#[derive(Copy, Clone, Eq, PartialEq)]
-#[repr(C)]
-pub struct Fr(U256);
-
-const R : U256 = U256([0x8944D45FD166C971, 0xF77965C4DFD30734, 0xFFFFFFFFFFFFFFFF, 0x1FFFFFFFFFFFFFF]); // |Fr|
-
-const ZR_SQUARED : Fr = Fr(U256([0x32A1CB0B0D0DF74A, 0x7FE44146DFCFDAF8, 0xAF59BBB1E8ACC494, 0x1A6134EBBFC821])); // (2^256)^2 mod |Fr|
-const FRINV : u64 = 0xCD27F41CB1C5286F; // (-1/|Fr|) mod 2^64
-const R_ONE : Fr = Fr(U256([0x5D95D0174C9B4780, 0x434D1D90167C65BB, 4, 0])); // = 2^256 mod |Fr|
-const ZR_CUBED : Fr = Fr(U256([0x4BC368544B1323FA, 0xACA9EEEE6129D3CC, 0xA005B5D44D4502BD, 0x11ECFA1EAC284DF]));
-
-impl PartialOrd for Fr {
-    fn partial_cmp(&self, other: &Fr) -> Option<Ordering> {
-        U256::partial_cmp(&self.0, &other.0)
-    }
-}
-
-impl Ord for Fr {
-    fn cmp(&self, other: &Fr) -> Ordering {
-        U256::cmp(&self.0, &other.0)
-    }
-}
-/* --- */
-
-// -------------------------------------------------------------------
-
-impl Fr {
-    pub fn zero() -> Fr {
-        Fr(U256::zero())
-    }
-
-    pub fn one() -> Fr {
-        R_ONE
-    }
-
-    pub fn invert(self) -> Fr {
-        let mut tmp = self;
-        U256::invert_mod(&mut tmp.0, &R);
-        ZR_CUBED * tmp
-    }
-}
-
-impl fmt::Display for Fr {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let FrUnscaled(tmp) = FrUnscaled::from(*self);
-        write!(f, "Fr({})", tmp.nbr_str())
-    }
-}
-
-struct FrUnscaled(U256);
-
-impl From<i64> for FrUnscaled {
-    fn from(x : i64) -> FrUnscaled {
-        if x >= 0 {
-            let z = U256([x as u64, 0, 0, 0]);
-            FrUnscaled(z)
-        } else {
-            let tmp = [(-x) as u64, 0, 0, 0];
-            let mut tmp2 = R.0;
-            sub_noborrow(&mut tmp2, &tmp);
-            let z = U256(tmp2);
-            FrUnscaled(z)
-        }
-    }
-}
-
-impl From<FrUnscaled> for Fr {
-    fn from(x : FrUnscaled) -> Fr {
-        let FrUnscaled(z) = x;
-        ZR_SQUARED * Fr(z)
-    }
-}
-
-impl From<i64> for Fr {
-    fn from(x : i64) -> Fr {
-        Fr::from(FrUnscaled::from(x))
-    }
-}
-
-impl Add<Fr> for Fr {
-    type Output = Fr;
-    fn add(self, other: Fr) -> Fr {
-        let mut tmp = self;
-        U256::add_mod(&mut tmp.0, &other.0, &R);
-        tmp
-    }
-}
-
-impl Add<i64> for Fr {
-    type Output = Fr;
-    fn add(self, other: i64) -> Fr {
-        self + Fr::from(other)
-    }
-}
-
-impl Add<Fr> for i64 {
-    type Output = Fr;
-    fn add(self, other: Fr) -> Fr {
-        Fr::from(self) + other
-    }
-}
-
-impl AddAssign<Fr> for Fr {
-    fn add_assign(&mut self, other: Fr) {
-        U256::add_mod(&mut self.0, &other.0, &R);
-    }
-}
-
-impl AddAssign<i64> for Fr {
-    fn add_assign(&mut self, other: i64) {
-        *self += Fr::from(other);
-    }
-}
-
-impl Sub<Fr> for Fr {
-    type Output = Fr;
-    fn sub(self, other: Fr) -> Fr {
-        let mut tmp = self;
-        U256::sub_mod(&mut tmp.0, &other.0, &R);
-        tmp
-    }
-}
-
-impl Sub<i64> for Fr {
-    type Output = Fr;
-    fn sub(self, other: i64) -> Fr {
-        self - Fr::from(other)
-    }
-}
-
-impl Sub<Fr> for i64 {
-    type Output = Fr;
-    fn sub(self, other: Fr) -> Fr {
-        Fr::from(self) - other
-    }
-}
-
-impl SubAssign<Fr> for Fr {
-    fn sub_assign(&mut self, other: Fr) {
-        U256::sub_mod(&mut self.0, &other.0, &R);
-    }
-}
-
-impl SubAssign<i64> for Fr {
-    fn sub_assign(&mut self, other: i64) {
-        *self -= Fr::from(other);
-    }
-}
-
-impl Neg for Fr {
-    type Output = Fr;
-    fn neg(self) -> Fr {
-        let mut tmp = self;
-        U256::neg_mod(&mut tmp.0, &R);
-        tmp
-    }
-}
-
-impl Mul<Fr> for Fr {
-    type Output = Fr;
-    fn mul(self, other: Fr) -> Fr {
-        let mut tmp = self;
-        U256::mul_mod(&mut tmp.0, &other.0, &R, FRINV);
-        tmp
-    }
-}
-
-impl Mul<i64> for Fr {
-    type Output = Fr;
-    fn mul(self, other: i64) -> Fr {
-        self * Fr::from(other)
-    }
-}
-
-impl Mul<Fr> for i64 {
-    type Output = Fr;
-    fn mul(self, other: Fr) -> Fr {
-        other * self
-    }
-}
-
-impl MulAssign<Fr> for Fr {
-    fn mul_assign(&mut self, other: Fr) {
-        U256::mul_mod(&mut self.0, &other.0, &R, FRINV);
-    }
-}
-
-impl MulAssign<i64> for Fr {
-    fn mul_assign(&mut self, other: i64) {
-        *self *= Fr::from(other);
-    }
-}
-
-impl Div<Fr> for Fr {
-    type Output = Fr;
-    fn div(self, other: Fr) -> Fr {
-        self * Fr::invert(other)
-    }
-}
-
-impl Div<i64> for Fr {
-    type Output = Fr;
-    fn div(self, other: i64) -> Fr {
-        self / Fr::from(other)
-    }
-}
-
-impl Div<Fr> for i64 {
-    type Output = Fr;
-    fn div(self, other: Fr) -> Fr {
-        if self == 1 {
-            Fr::invert(other)
-        } else {
-            Fr::from(self) / other
-        }
-    }
-}
-
-impl DivAssign<Fr> for Fr {
-    fn div_assign(&mut self, other: Fr) {
-        *self *= Fr::invert(other);
-    }
-}
-
-impl DivAssign<i64> for Fr {
-    fn div_assign(&mut self, other: i64) {
-        *self /= Fr::from(other);
-    }
-}
-
-// -----------------------------------------------------------------
-// Fq is the field in which the curve is computed - coords are all elements of Fq
-// In Elliptic curve point operations these coordinates are converted to Fq51 representation
-//
-// Type Fq is for working directly in the field Fq, using fast Montgomery reduction
-// for modular multiply. As such, Fq is scaled by the Q_ONE value. Coordinate values
-// Fq51 must come from unscaled Fq values, and for that we have FqUnscaled types to 
-// help avoid the overhead of scaling / de-scaling. 
-
-#[derive(Copy, Clone, Eq, PartialEq)]
-#[repr(C)]
-pub struct Fq(U256);
-
-const Q : U256 = U256([0xFFFFFFFFFFFFFFF7, 0xFFFFFFFFFFFFFFFF, 0xFFFFFFFFFFFFFFFF, 0x7FFFFFFFFFFFFFF]); // |Fq|
-
-const ZQ_SQUARED : Fq = Fq(U256([0x014400, 0x00, 0x00, 0x00])); // (2^256)^2 mod |Fq|
-const FQINV : u64 = 0x8E38E38E38E38E39; // (-1/|Fq|) mod 2^64
-const Q_ONE : Fq = Fq(U256([0x0120, 0x00, 0x00, 0x00])); // = 2^256 mod |Fq|
-const ZQ_CUBED : Fq = Fq(U256([0x016C8000, 0x00, 0x00, 0x00])); // = (2^256)^3 mod |Fq|
-
-impl PartialOrd for Fq {
-    fn partial_cmp(&self, other: &Fq) -> Option<Ordering> {
-        U256::partial_cmp(&self.0, &other.0)
-    }
-}
-
-impl Ord for Fq {
-    fn cmp(&self, other: &Fq) -> Ordering {
-        U256::cmp(&self.0, &other.0)
-    }
-}
-
-impl Fq {
-    pub fn zero() -> Fq {
-        Fq(U256::zero())
-    }
-
-    pub fn one() -> Fq {
-        Q_ONE
-    }
-
-    pub fn invert(self) -> Fq {
-        let mut tmp = self;
-        U256::invert_mod(&mut tmp.0, &Q);
-        ZQ_CUBED * tmp
-    }
-}
-
-impl fmt::Display for Fq {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let FqUnscaled(tmp) = FqUnscaled::from(*self);
-        write!(f, "Fq({})", tmp.nbr_str())
-    }
-}
-
-struct FqUnscaled(U256);
-
-impl From<i64> for FqUnscaled {
-    fn from(x : i64) -> FqUnscaled {
-        if x >= 0 {
-            let z = U256([x as u64, 0, 0, 0]);
-            FqUnscaled(z)
-        } else {
-            let tmp = [(-x) as u64, 0, 0, 0];
-            let mut tmp2 = Q.0;
-            sub_noborrow(&mut tmp2, &tmp);
-            let z = U256(tmp2);
-            FqUnscaled(z)
-        }
-    }
-}
-
-impl From<FqUnscaled> for Fq {
-    fn from(x : FqUnscaled) -> Fq {
-        let FqUnscaled(z) = x;
-        ZQ_SQUARED * Fq(z)
-    }
-}
-
-impl From<i64> for Fq {
-    fn from(x : i64) -> Fq {
-        Fq::from(FqUnscaled::from(x))
-    }
-}
-
-impl Add<Fq> for Fq {
-    type Output = Fq;
-    fn add(self, other: Fq) -> Fq {
-        let mut tmp = self;
-        U256::add_mod(&mut tmp.0, &other.0, &Q);
-        tmp
-    }
-}
-
-impl Add<i64> for Fq {
-    type Output = Fq;
-    fn add(self, other: i64) -> Fq {
-        self + Fq::from(other)
-    }
-}
-
-impl Add<Fq> for i64 {
-    type Output = Fq;
-    fn add(self, other: Fq) -> Fq {
-        Fq::from(self) + other
-    }
-}
-
-impl Sub<Fq> for Fq {
-    type Output = Fq;
-    fn sub(self, other: Fq) -> Fq {
-        let mut tmp = self;
-        U256::sub_mod(&mut tmp.0, &other.0, &Q);
-        tmp
-    }
-}
-
-impl Sub<i64> for Fq {
-    type Output = Fq;
-    fn sub(self, other: i64) -> Fq {
-        self - Fq::from(other)
-    }
-}
-
-impl Sub<Fq> for i64 {
-    type Output = Fq;
-    fn sub(self, other: Fq) -> Fq {
-        Fq::from(self) - other
-    }
-}
-
-impl Neg for Fq {
-    type Output = Fq;
-    fn neg(self) -> Fq {
-        let mut tmp = self;
-        U256::neg_mod(&mut tmp.0, &Q);
-        tmp
-    }
-}
-
-impl Mul<Fq> for Fq {
-    type Output = Fq;
-    fn mul(self, other: Fq) -> Fq {
-        let mut tmp = self;
-        U256::mul_mod(&mut tmp.0, &other.0, &Q, FQINV);
-        tmp
-    }
-}
-
-impl Mul<i64> for Fq {
-    type Output = Fq;
-    fn mul(self, other: i64) -> Fq {
-        self * Fq::from(other)
-    }
-}
-
-impl Mul<Fq> for i64 {
-    type Output = Fq;
-    fn mul(self, other: Fq) -> Fq {
-        other * self
-    }
-}
-
-impl Div<Fq> for Fq {
-    type Output = Fq;
-    fn div(self, other: Fq) -> Fq {
-        self * Fq::invert(other)
-    }
-}
-
-impl Div<i64> for Fq {
-    type Output = Fq;
-    fn div(self, other: i64) -> Fq {
-        self / Fq::from(other)
-    }
-}
-
-impl Div<Fq> for i64 {
-    type Output = Fq;
-    fn div(self, other: Fq) -> Fq {
-        if self == 1 {
-            Fq::invert(other)
-        } else {
-            Fq::from(self) / other
-        }
-    }
-}
-
-// -----------------------------------------------------------------
-// field Fq51 is Fq broken into 51-bit frames 
-// in little-endian order, of i64 (vs u64 in Fq)
-// 47 bits in last frame
-
-#[derive(Copy, Clone)]
-#[repr(C)]
-pub struct Fq51([i64;5]);
-
-const FQ51_0 : Fq51 = Fq51([0;5]);
-const FQ51_1 : Fq51 = Fq51([1,0,0,0,0]);
-
-impl Fq51 {
-    pub fn zero() -> Fq51 {
-        FQ51_0
-    }
-
-    pub fn one() -> Fq51 {
-        FQ51_1
-    }
-
-    pub fn is_zero(&self) -> bool {
-        // faster than test == FQ51_0
-        // avoid scr(FQ51_0);
-        let mut tmp = *self;
-        scr(&mut tmp);
-        tmp.0 == FQ51_0.0
-    }
-
-    pub fn is_one(&self) -> bool {
-        // faster than test == FQ51_1
-        // avoid scr(FQ51_1);
-        let mut tmp = *self;
-        scr(&mut tmp);
-        tmp.0 == FQ51_1.0
-    }
-    
-    pub fn is_odd(&self) -> bool {
-        (self.0[0] & 1) != 0
-    }
-
-    pub fn sqr(self) -> Fq51 {
-        let mut tmp = Fq51::zero();
-        gsqr(&self, &mut tmp);
-        tmp
-    }
-
-    fn nbr_str(&self) -> String {
-        let mut y: U256 = U256::zero();
-        elt_to_bin(self, &mut y);
-        let U256(yv) = y;
-        basic_nbr_str(&yv)
-    }
-}
-
-impl From<i64> for Fq51 {
-    fn from(x: i64) -> Fq51 {
-        Fq51::from(FqUnscaled::from(x))
-    }
-}
-
-impl From<FqUnscaled> for Fq51 {
-    fn from(x : FqUnscaled) -> Fq51 {
-        let mut tmp = Fq51::zero();
-        bin_to_elt(&x.0, &mut tmp);
-        tmp
-    }
-}
-
-impl From<Fq> for Fq51 {
-    fn from(x : Fq) -> Fq51 {
-        Fq51::from(FqUnscaled::from(x))
-    }
-}
-
-impl From<Fq51> for FqUnscaled {
-    fn from(x : Fq51) -> FqUnscaled {
-        let mut tmp = U256::zero();
-        elt_to_bin(&x, &mut tmp);
-        FqUnscaled(tmp)
-    }
-}
-
-impl fmt::Debug for Fq51 {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "Fq51([{:016x}, {:016x}, {:016x}, {:016x}, {:016x}])",
-          self.0[0] as u64, 
-          self.0[1] as u64,
-          self.0[2] as u64,
-          self.0[3] as u64,
-          self.0[4] as u64)
-    }
-}
-
-impl fmt::Display for Fq51 {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "Fq51({})", self.nbr_str())
-    }
-}
-
-impl Add<Fq51> for Fq51 {
-    type Output = Fq51;
-    fn add(self, other: Fq51) -> Fq51 {
-        let mut dst = Fq51::zero();
-        for i in 0..5 {
-            dst.0[i] = self.0[i] + other.0[i];
-        }
-        dst
-    }
-}
-
-impl Add<i64> for Fq51 {
-    type Output = Fq51;
-    fn add(self, other: i64) -> Fq51 {
-        let mut dst = self;
-        dst.0[0] += other;
-        dst
-    }
-}
-
-impl Add<Fq51> for i64 {
-    type Output = Fq51;
-    fn add(self, other: Fq51) -> Fq51 {
-        let mut dst = other;
-        dst.0[0] += self;
-        dst
-    }
-}
-
-impl AddAssign<Fq51> for Fq51 {
-    fn add_assign(&mut self, other: Fq51) {
-        for i in 0..5 {
-            self.0[i] += other.0[i];
-        }
-    }
-}
-
-impl AddAssign<i64> for Fq51 {
-    fn add_assign(&mut self, other: i64) {
-        self.0[0] += other;
-    }
-}
-
-impl Sub<Fq51> for Fq51 {
-    type Output = Fq51;
-    fn sub(self, other: Fq51) -> Fq51 {
-        let mut dst = Fq51::zero();
-        for i in 0..5 {
-            dst.0[i] = self.0[i] - other.0[i];
-        }
-        dst
-    }
-}
-
-impl Sub<i64> for Fq51 {
-    type Output = Fq51;
-    fn sub(self, other: i64) -> Fq51 {
-        let mut dst = self;
-        dst.0[0] -= other;
-        dst
-    }
-}
-
-impl Sub<Fq51> for i64 {
-    type Output = Fq51;
-    fn sub(self, other: Fq51) -> Fq51 {
-        Fq51::from(self) - other
-    }
-}
-
-impl SubAssign<Fq51> for Fq51 {
-    fn sub_assign(&mut self, other: Fq51) {
-        for i in 0..5 {
-            self.0[i] -= other.0[i];
-        }
-    }
-}
-
-impl SubAssign<i64> for Fq51 {
-    fn sub_assign(&mut self, other: i64) {
-        self.0[0] -= other;
-    }
-}
-
-impl Neg for Fq51 {
-    type Output = Fq51;
-    fn neg(self) -> Fq51 {
-        let mut dst = Fq51::zero();
-        for i in 0..5 {
-            dst.0[i] = -self.0[i];
-        }
-        dst
-    }
-}
-
-impl Mul<Fq51> for Fq51 {
-    type Output = Fq51;
-    fn mul(self, other: Fq51) -> Fq51 {
-        let mut dst = Fq51::zero();
-        gmul(&self, &other, &mut dst);
-        dst
-    }
-}
-
-impl MulAssign<Fq51> for Fq51 {
-    fn mul_assign(&mut self, other: Fq51) {
-        let tmp = *self;
-         gmul(&tmp, &other, self);
-    }
-}
-
-impl Mul<Fq51> for i64 {
-    type Output = Fq51;
-    fn mul(self, other: Fq51) -> Fq51 {
-        let mut dst = other;
-        gmuli(&mut dst, self);
-        dst
-    }
-}
-
-impl Mul<i64> for Fq51 {
-    type Output = Fq51;
-    fn mul(self, other: i64) -> Fq51 {
-        let mut dst = self;
-        gmuli(&mut dst, other);
-        dst
-    }
-}
-
-impl MulAssign<i64> for Fq51 {
-    fn mul_assign(&mut self, other: i64) {
-        gmuli(self, other);
-    }
-}
-
-impl Div<Fq51> for Fq51 {
-    type Output = Fq51;
-    fn div(self, x : Fq51) -> Fq51 {
-        let mut tmp = x;
-        ginv(&mut tmp);
-        self * tmp
-    }
-}
-
-impl PartialEq for Fq51 {
-    fn eq(&self, other : &Fq51) -> bool {
-        let mut a = *self;
-        let mut b = *other;
-        scr(&mut a);
-        scr(&mut b);
-        a.0 == b.0
-    }
-}
-// ---------------------------------------------------------
-// Group primitive operators
-
-fn gadd(x: &Fq51, y: &Fq51, w: &mut Fq51) {
-    w.0[0] = x.0[0] + y.0[0];
-    w.0[1] = x.0[1] + y.0[1];
-    w.0[2] = x.0[2] + y.0[2];
-    w.0[3] = x.0[3] + y.0[3];
-    w.0[4] = x.0[4] + y.0[4];
-}
-
-fn gsub(x: &Fq51, y: &Fq51, w: &mut Fq51) {
-    w.0[0] = x.0[0] - y.0[0];
-    w.0[1] = x.0[1] - y.0[1];
-    w.0[2] = x.0[2] - y.0[2];
-    w.0[3] = x.0[3] - y.0[3];
-    w.0[4] = x.0[4] - y.0[4];
-}
-
-fn gdec(x: &Fq51, w: &mut Fq51) {
-    w.0[0] -= x.0[0];
-    w.0[1] -= x.0[1];
-    w.0[2] -= x.0[2];
-    w.0[3] -= x.0[3];
-    w.0[4] -= x.0[4];
-}
-
-fn gneg(w: &Fq51, x: &mut Fq51) {
-    x.0[0] = -w.0[0];
-    x.0[1] = -w.0[1];
-    x.0[2] = -w.0[2];
-    x.0[3] = -w.0[3];
-    x.0[4] = -w.0[4];
-}
-
-// w*=2
-fn gmul2(w: &mut Fq51) {
-    w.0[0] *= 2;
-    w.0[1] *= 2;
-    w.0[2] *= 2;
-    w.0[3] *= 2;
-    w.0[4] *= 2;
-}
-
-// w-=2*x
-fn gsb2(x: &Fq51, w: &mut Fq51) {
-    w.0[0] -= 2*x.0[0];
-    w.0[1] -= 2*x.0[1];
-    w.0[2] -= 2*x.0[2];
-    w.0[3] -= 2*x.0[3];
-    w.0[4] -= 2*x.0[4];
-}
-
-// reduce w - Short Coefficient Reduction
-fn scr(w: &mut Fq51) {
-    loop {
-        let w0 = w.0[0];
-        let t0 = w0 & BOT_51_BITS;
-    
-        let t1 = w.0[1] + (w0 >> 51);
-        w.0[1] = t1 & BOT_51_BITS;
-
-        let t2 = w.0[2] + (t1 >> 51);
-        w.0[2] = t2 & BOT_51_BITS;
-
-        let t3 = w.0[3] + (t2 >> 51);
-        w.0[3] = t3 & BOT_51_BITS;
-
-        let t4 = w.0[4] + (t3 >> 51);
-        w.0[4] = t4 & BOT_47_BITS;
-        w.0[0] = t0 + 9*(t4 >> 47);
-
-        if (w.0[0] >> 51) == 0 { break; }
-    } 
-}
-
-// multiply w by a constant, w*=i
-
-fn gmuli(w: &mut Fq51, i: i64) {
-    let ii = i as i128;
-    let t0 = (w.0[0] as i128) * ii;
-    w.0[0] = (t0 as i64) & BOT_51_BITS;
-
-    let t1 = (w.0[1] as i128) * ii + (t0 >> 51);
-    w.0[1] = (t1 as i64) & BOT_51_BITS;
-
-    let t2 = (w.0[2] as i128) * ii + (t1 >> 51);
-    w.0[2] = (t2 as i64) & BOT_51_BITS;
-
-    let t3 = (w.0[3] as i128) * ii + (t2 >> 51);
-    w.0[3] = (t3 as i64) & BOT_51_BITS;
-
-    let t4 = (w.0[4] as i128) * ii + (t3 >> 51);
-    w.0[4] = (t4 as i64) & BOT_47_BITS;
-    w.0[0] += (9 * (t4 >> 47)) as i64;
-}
-
-// z=x^2
-
-fn gsqr(x: &Fq51, z: &mut Fq51) {
-    let t4 = 2*((x.0[0] as i128) * (x.0[4] as i128) +
-                (x.0[1] as i128) * (x.0[3] as i128)) +
-            (x.0[2] as i128) * (x.0[2] as i128);
-    z.0[4] = (t4 as i64) & BOT_47_BITS;
-
-    let t0 = (x.0[0] as i128)*(x.0[0] as i128) +
-                 288*((x.0[1] as i128)*(x.0[4] as i128) +
-                      (x.0[2] as i128)*(x.0[3] as i128)) +
-                   9*(t4 >> 47);
-    z.0[0] = (t0 as i64) & BOT_51_BITS;
-    
-    let t1 = 2*(x.0[0] as i128)*(x.0[1] as i128) +
-         288*(x.0[2] as i128)*(x.0[4] as i128) +
-         144*(x.0[3] as i128)*(x.0[3] as i128) +
-         (t0 >> 51);
-    z.0[1] = (t1 as i64) & BOT_51_BITS;
-
-    let t2 = (x.0[1] as i128)*(x.0[1] as i128) +
-         2*(x.0[0] as i128)*(x.0[2] as i128) +
-         288*(x.0[3] as i128)*(x.0[4] as i128) +
-         (t1 >> 51);
-    z.0[2] = (t2 as i64) & BOT_51_BITS;
-
-    let t3 = 144*(x.0[4] as i128)*(x.0[4] as i128) +
-           2*((x.0[0] as i128)*(x.0[3] as i128) +
-              (x.0[1] as i128)*(x.0[2] as i128)) +
-         (t2 >> 51);
-    z.0[3] = (t3 as i64) & BOT_51_BITS;
-
-    let t4 = (z.0[4] as i128) + (t3 >> 51);
-    z.0[4] = (t4 as i64) & BOT_47_BITS;
-    z.0[0] += (9*(t4 >> 47)) as i64;
-}
-
-fn gmul(x: &Fq51, y: &Fq51, z: &mut Fq51) {
-	// 5M + 4A
-    let t4 = (x.0[0] as i128)*(y.0[4] as i128) +
-        (x.0[4] as i128)*(y.0[0] as i128) +
-        (x.0[1] as i128)*(y.0[3] as i128) +
-        (x.0[3] as i128)*(y.0[1] as i128) +
-        (x.0[2] as i128)*(y.0[2] as i128);
-    z.0[4] = (t4 as i64) & BOT_47_BITS;
-    
-	// 7M + 5A
-    let t0 = (x.0[0] as i128)*(y.0[0] as i128) +
-        144*((x.0[1] as i128)*(y.0[4] as i128) +
-             (x.0[4] as i128)*(y.0[1] as i128) +
-             (x.0[2] as i128)*(y.0[3] as i128) +
-             (x.0[3] as i128)*(y.0[2] as i128)) +
-        9*(t4 >> 47);
-    z.0[0] = (t0 as i64) & BOT_51_BITS;
-
-	// 6M + 5A
-    let t1 = (x.0[0] as i128)*(y.0[1] as i128) +
-         (x.0[1] as i128)*(y.0[0] as i128) +
-         144*((x.0[3] as i128)*(y.0[3] as i128) +
-              (x.0[2] as i128)*(y.0[4] as i128) +
-              (x.0[4] as i128)*(y.0[2] as i128)) +
-         (t0 >> 51);
-    z.0[1] = (t1 as i64) & BOT_51_BITS;
-    
-    // 6M + 5A
-    let t2 = (x.0[1] as i128)*(y.0[1] as i128) +
-         (x.0[0] as i128)*(y.0[2] as i128) +
-         (x.0[2] as i128)*(y.0[0] as i128) +
-         144*((x.0[3] as i128)*(y.0[4] as i128) +
-              (x.0[4] as i128)*(y.0[3] as i128)) +
-         (t1 >> 51);
-    z.0[2] = (t2 as i64) & BOT_51_BITS;
-    
-    // 6M + 5A
-    let t3 = 144*((x.0[4] as i128)*(y.0[4] as i128)) +
-         (x.0[0] as i128)*(y.0[3] as i128) +
-         (x.0[3] as i128)*(y.0[0] as i128) +
-         (x.0[1] as i128)*(y.0[2] as i128) +
-         (x.0[2] as i128)*(y.0[1] as i128) +
-         (t2 >> 51);
-    z.0[3] = (t3 as i64) & BOT_51_BITS;
-
-    // -------- to this point = 30M + 24A => this clocks as faster than Granger's method for Curve1174
-    let t4 = (z.0[4] as i128) + (t3 >> 51);
-    z.0[4] = (t4 as i64) & BOT_47_BITS;
-    z.0[0] += (9*(t4 >> 47)) as i64;
-}
-
-// Inverse x = 1/x = x^(p-2) mod p
-// the exponent (p-2) = "07FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF5"
-// (61 F's)
-
-fn ginv(x: &mut Fq51) {
-    let mut w  = FQ51_0;
-    let mut t1 = FQ51_0;
-    let mut t2;
-    let mut x3 = FQ51_0;
-    let mut x5 = FQ51_0;
-    
-    // --------------------------------------
-    // 265*M
-    gsqr(x, &mut w);      // w = x^2
-    gmul(x, &w, &mut x3);   // x3 = x^3 = x^(2^2-1)
-    gmul(&w, &x3, &mut x5);  // x5 = x^5
-
-    gsqr(&x3, &mut w);
-    gsqr(&w, &mut t1);     // t1 = x^(2^4-4)
-    gmul(&x3, &t1, &mut w);  // w = x^(2^4-1)
-
-    gsqr(&w, &mut t1);
-    gsqr(&t1, &mut w);     // w = x^(2^6-4)
-    gmul(&x3, &w, &mut t1);  // t1 = x^(2^6-1)
-    t2 = t1;
-    for _ in 0..3 {
-	gsqr(&t1, &mut w);
-	gsqr(&w, &mut t1);
-    }
-    gmul(&t1, &t2, &mut w);  // w = x^(2^12-1)
-    
-    gsqr(&w, &mut t1);
-    gsqr(&t1, &mut w);     // w = x^(2^14-4)
-    gmul(&x3, &w, &mut t1);  // t1 = x^(2^14-1)
-    t2 = t1;
-    for _ in 0..7 {
-	gsqr(&t1, &mut w);
-	gsqr(&w, &mut t1);
-    }
-    gmul(&t1, &t2, &mut w);  // w = x^(2^28-1)
-    
-    gsqr(&w, &mut t1);
-    gsqr(&t1, &mut w);     // w = x^(2^30-4)
-    gmul(&x3, &w, &mut t1);  // t1 = x^(2^30-1)
-    t2 = t1;
-    for _ in 0..15 {
-	gsqr(&t1, &mut w);
-	gsqr(&w, &mut t1);
-    }
-    gmul(&t1, &t2, &mut w);  // w = x^(2^60-1)
-    
-    t2 = w;
-    for _ in 0..30 {
-	gsqr(&w, &mut t1);
-	gsqr(&t1, &mut w);
-    }
-    gmul(&w, &t2, &mut t1);  // t1 = x^(2^120-1)
-    
-    gsqr(&t1, &mut w);
-    gsqr(&w, &mut t1);     // t1 = x^(2^122 - 4)
-    gmul(&x3, &t1, &mut w);  // w = x^(2^122-1)
-    t2 = w;
-    for _ in 0..61 {
-	gsqr(&w, &mut t1);
-	gsqr(&t1, &mut w);
-    }
-    gmul(&w, &t2, &mut t1);  // t1 = x^(2^244-1)
-    
-    gsqr(&t1, &mut w);     // w = x^(2^245-2)
-    gmul(&x, &w, &mut t1);   // t1 = x^(2^245-1)
-    gsqr(&t1, &mut w);      
-    gsqr(&w, &mut t1);     // t1 = x^(2^247-4)
-    gmul(&x3, &t1, &mut w);  // w = x^(2^247-1)
-    
-    gsqr(&w, &mut t1);
-    gsqr(&t1, &mut w);
-    gsqr(&w, &mut t1);
-    gsqr(&t1, &mut w);     // w = x^(2^251-16)
-    gmul(&x5, &w, x);   // x = x^(2^251-11)
-}
-
-fn gdec2(x: &mut Fq51) {
-    x.0[0] -= 2;
-}
-
-fn gsqrt(x: Fq51) -> Option<Fq51> {
-    // we need to perform (x^((q+1)/4) mod q)
-    // for (q + 1)/4 = 0x01FF__FFFF__FFFF_FFFF__FFFF_FFFF_FFFF_FFFF__FFFF_FFFF_FFFF_FFFF_FFFF_FFFF_FFFF_FFFE
-    //               = 2^(2*(248-1))
-    // At end we will verify: sqrt(x)^2 mod q == x, bypassing usual check for Legendre symbol = +1.
-
-    let mut w  = FQ51_0;
-    let mut t1 = FQ51_0;
-    let mut t2 = FQ51_0;
-
-    let mut t8   = FQ51_0;
-    let mut t16  = FQ51_0;
-    let mut t32  = FQ51_0;
-    let mut t64  = FQ51_0;
-
-    // --------------------------------------
-    // 260*M
-
-    gsqr(&x, &mut w);         // w = x^2
-    gmul(&x, &w, &mut t1);    // t1 = x^3
-    gsqr(&t1, &mut w);        // w = x^6
-    gsqr(&w, &mut t2);        // t2 = x^12
-    gmul(&t1, &t2, &mut w);   // w = x^15 = x^(2^4-1)
-
-    t2 = w;
-    for _ in 0 .. 2 {
-        gsqr(&w, &mut t1);
-        gsqr(&t1, &mut w);
-    }
-    gmul(&t2, &w, &mut t8);  // t8 = x^(2^8-1)
-
-    w = t8;
-    for _ in 0 .. 4 {
-        gsqr(&w, &mut t1);
-        gsqr(&t1, &mut w);
-    }
-    gmul(&t8, &w, &mut t16);  // t16 = x^(2^16-1)
-
-    w = t16;
-    for _ in 0 .. 8 {
-        gsqr(&w, &mut t1);
-        gsqr(&t1, &mut w);
-    }
-    gmul(&t16, &w, &mut t32); // t32 = x^(2^32-1)
-
-    w = t32;
-    for _ in 0 .. 16 {
-        gsqr(&w, &mut t1);
-        gsqr(&t1, &mut w);
-    }
-    gmul(&t32, &w, &mut t64); // t64 = x^(2^64-1)
-
-    w = t64;
-    for _ in 0 .. 32 {
-        gsqr(&w, &mut t1);
-        gsqr(&t1, &mut w);
-    }
-    gmul(&t64, &w, &mut t1);  // t1 = x^(2^128-1)
-
-    for _ in 0 .. 32 {
-        gsqr(&t1, &mut w);
-        gsqr(&w, &mut t1);
-    }
-    gmul(&t64, &t1, &mut w); // w = x^(2^192-1)
-
-    for _ in 0 .. 16 {
-        gsqr(&w, &mut t1);
-        gsqr(&t1, &mut w);
-    }
-    gmul(&t32, &w, &mut t1); // t1 = x^(2^224-1)
-
-    for _ in 0 .. 8 {
-        gsqr(&t1, &mut w);
-        gsqr(&w, &mut t1);
-    }
-    gmul(&t16, &t1, &mut w); // w = x^(2^240-1)
-
-    for _ in 0 .. 4 {
-        gsqr(&w, &mut t1);
-        gsqr(&t1, &mut w);
-    }
-    gmul(&t8, &w, &mut t1); // t1 = x^(2^248-1)
-
-    gsqr(&t1, &mut w);      // w = x^(2*(248-1)) =? sqrt(x)
-
-    gsqr(&w, &mut t1);      // t1 = sqrt(x)^2 =? x
-    if t1 == x { Some(w) } else { None }
-}
-
-// -------------------------------------------------------------------------
-// Point Structure
-
-#[derive(Copy, Clone, Debug)]
-#[repr(C)]
-pub struct ECp { x: Fq51,
-                 y: Fq51,
-                 z: Fq51,
-                 t: Fq51 }
-
-const PT_INF: ECp  = ECp { x: FQ51_0,
-                           y: FQ51_1,
-                           z: FQ51_1,
-                           t: FQ51_0 };
-
-impl ECp {
-    pub fn inf() -> ECp {
-        PT_INF
-    }
-
-    pub fn is_inf(&self) -> bool {
-        self.x.is_zero()
-    }
-
-    pub fn is_affine(&self) -> bool {
-        self.z.is_one()
-    }
-
-    pub fn new(x: Fq, y: Fq) -> ECp {
-        let fq51_x = Fq51::from(x);
-        let fq51_y = Fq51::from(y);
-        ECp::from_xy(fq51_x, fq51_y)
-    }
-
-    pub fn from_xy(x: Fq51, y: Fq51) -> ECp {
-        ECp { x: x,
-              y: y,
-              z: Fq51::one(),
-              t: x * y }
-    }
-}
-
-impl fmt::Display for ECp {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "ECp {{\n x: {},\n y: {},\n z: {},\n t: {} }}", 
-            self.x, self.y, self.z, self.t)
-    }
-}
-
-impl Add<ECp> for ECp {
-    type Output = ECp;
-    fn add(self, other: ECp) -> ECp {
-        let mut tmp = PT_INF;
-        add_proj(&self, &other, &mut tmp);
-        tmp
-    }
-}
-
-impl AddAssign<ECp> for ECp {
-    fn add_assign(&mut self, other: ECp) {
-        let tmp = *self;
-        add_proj(&tmp, &other, self);
-    }
-}
-
-impl Neg for ECp {
-    type Output = ECp;
-    fn neg(self) -> ECp {
-        let mut tmp = ECp::inf();
-        tmp.x = -self.x;
-        tmp.y =  self.y;
-        tmp.z =  self.z;
-        tmp.t = -self.t;
-        tmp
-    }
-}
-
-impl Sub<ECp> for ECp {
-    type Output = ECp;
-    fn sub(self, other: ECp) -> ECp {
-        self + (- other)
-    }
-}
-
-impl SubAssign<ECp> for ECp {
-    fn sub_assign(&mut self, other: ECp) {
-        *self += -other;
-    }
-}
-
-impl Mul<Fr> for ECp {
-    type Output = ECp;
-    fn mul(self, other: Fr) -> ECp {
-        let wv = WinVec::from(other);
-        let mut tmp = self;
-        ecp_mul(&wv, &mut tmp);
-        tmp
-    }
-}
-
-impl Mul<ECp> for Fr {
-    type Output = ECp;
-    fn mul(self, other: ECp) -> ECp {
-        other * self
-    }
-}
-
-impl MulAssign<Fr> for ECp {
-    fn mul_assign(&mut self, other: Fr) {
-        let wv = WinVec::from(other);
-        ecp_mul(&wv, self);
-    }
-}
-
-impl Mul<i64> for ECp {
-    type Output = ECp;
-    fn mul(self, other: i64) -> ECp {
-        match other {
-            0 => ECp::inf(),
-            1 => self,
-            _ => {
-                    let wv = WinVec::from(other);
-                    let mut tmp = self;
-                    ecp_mul(&wv, &mut tmp);
-                    tmp
-            }
-        }
-    }
-}
-
-impl Mul<ECp> for i64 {
-    type Output = ECp;
-    fn mul(self, other: ECp) -> ECp {
-        other * self
-    }
-}
-
-impl MulAssign<i64> for ECp {
-    fn mul_assign(&mut self, other: i64) {
-        match other {
-            0 => *self = ECp::inf(),
-            1 => (),
-            _ => {
-                    let wv = WinVec::from(other);
-                    ecp_mul(&wv, self);
-            }
-        }
-    }
-}
-
-impl Div<Fr> for ECp {
-    type Output = ECp;
-    fn div(self, other: Fr) -> ECp {
-        self * Fr::invert(other)
-    }
-}
-
-impl DivAssign<Fr> for ECp {
-    fn div_assign(&mut self, other: Fr) {
-        *self *= Fr::invert(other);
-    }
-}
-
-impl Div<i64> for ECp {
-    type Output = ECp;
-    fn div(self, other: i64) -> ECp {
-        self / Fr::from(other)
-    }
-}
-
-impl DivAssign<i64> for ECp {
-    fn div_assign(&mut self, other: i64) {
-        *self /= Fr::from(other);
-    }
-}
-
-// --------------------------------------------------------
-// Point Operators
-
-// P+=P
-
-fn double_1(pt: &mut ECp) {
-    let a = pt.x.sqr();
-    let b = pt.y.sqr();
-    let e = 2 * pt.t;
-    let g = a + b;
-    let f = g - 2;
-    let h = a - b;
-    pt.x = e * f;
-    pt.y = g * h;
-    pt.z = g * (g - 2);
-    pt.t = e * h;
- }
-
-fn double_2(pt: &mut ECp) {
-    let a = pt.x.sqr();
-    let b = pt.y.sqr();
-    let c = 2 * pt.z.sqr();
-    let g = a + b;
-    let h = a - b;
-    let f = g - c;
-    let e = (pt.x + pt.y).sqr() - g;
-    pt.x = e * f;
-    pt.y = g * h;
-    pt.z = f * g;
-}
-
-fn double_3(pt : &mut ECp) {
-    let a = pt.x.sqr();
-    let b = pt.y.sqr();
-    let c = 2 * pt.z.sqr();
-    let g = a + b;
-    let h = a - b;
-    let f = g - c;
-    let e = (pt.x + pt.y).sqr() - g;
-    pt.x = e * f;
-    pt.y = g * h;
-    pt.z = f * g;
-    pt.t = e * h;
-}
-
-//P+=Q;
-
-fn add_1(qpt: &ECp, ppt: &mut ECp) {
-    let a = ppt.x * qpt.x;
-    let b = ppt.y * qpt.y;
-    let c = ppt.t * qpt.t;
-    let f = ppt.z - c;  // reversed sign as d is negative
-    let g = ppt.z + c;
-    let h = b - a;
-    let c = ppt.x + ppt.y;
-    let d = qpt.x + qpt.y;
-    let e = (c * d) - a - b;
-    ppt.x = e * f;
-    ppt.y = g * h;
-    ppt.z = f * g;
-    ppt.t = e * h;
-}
-
-fn add_2(qpt: &ECp, ppt: &mut ECp) {
-    let a = ppt.x * qpt.x;
-    let b = ppt.y * qpt.y;
-    let c = ppt.t * qpt.t;
-    let d = ppt.z * qpt.z;
-    let f = d - c; // reversed sign as d is negative
-    let g = d + c;
-    let h = b - a;
-    let c = ppt.x + ppt.y;
-    let d = qpt.x + qpt.y;
-    let e = (c * d) - a - b;
-    ppt.x = e * f;
-    ppt.y = g * h;
-    ppt.z = f * g;
-}
-
-//P=0
-
-// Initialise P
-// incoming x,y coordinates -> ECp in projective coords
-
-fn init(x: &Fq51, y: &Fq51, pt: &mut ECp) {
-    (*pt).x = *x;
-    (*pt).y = *y;
-    (*pt).z = Fq51::one();
-    gmul(x, y, &mut pt.t);
-}
-
-// P=-Q
-
-fn ecp_neg(qpt: &ECp, ppt: &mut ECp) {
-    gneg(&qpt.x, &mut ppt.x);
-    ppt.y = qpt.y;
-    ppt.z = qpt.z;
-    gneg(&qpt.t, &mut ppt.t);
-}
-   
-// Make Affine
-
-fn norm(pt: &mut ECp) {
-    let mut w = pt.z;
-    ginv(&mut w);
-    // (*pt).z = Fq51::one();
-
-    let mut tmp = pt.x * w;
-    scr(&mut tmp);
-    pt.x = tmp;
-
-    let mut tmp = pt.y * w;
-    scr(&mut tmp);
-    pt.y = tmp;
-
-    let mut tmp = pt.z * w;
-    scr(&mut tmp);
-    pt.z = tmp;
-
-    let mut tmp = pt.t * w;
-    scr(&mut tmp);
-    pt.t = tmp;
-}
-
-// Precomputation
-
-fn precomp(pt: &ECp, wpts: &mut [ECp]) {
-    let mut tmp1 = *pt;
-    gmuli(&mut tmp1.t, CURVE_D);
-
-    let mut tmp2 = *pt;
-    double_1(&mut tmp2);
-    
-    let mut tmp3 = tmp2;
-    add_1(&tmp1, &mut tmp3);
-
-    let mut tmp4 = tmp2;
-    double_3(&mut tmp4);
-
-    let mut tmp5 = tmp4;
-    add_1(&tmp1, &mut tmp5);
-
-    let mut tmp6 = tmp3;
-    double_3(&mut tmp6);
-
-    let mut tmp7 = tmp6;
-    add_1(&tmp1, &mut tmp7);
-
-    let mut tmp8 = tmp4;
-    double_3(&mut tmp8);
-    
-    // premultiply t parameter by curve constant
-    tmp2.t *= CURVE_D;
-    tmp3.t *= CURVE_D;
-    tmp4.t *= CURVE_D;
-    tmp5.t *= CURVE_D;
-    tmp6.t *= CURVE_D;
-    tmp7.t *= CURVE_D;
-    tmp8.t *= CURVE_D;
-
-    wpts[0] = PT_INF;
-    wpts[1] = tmp1;
-    wpts[2] = tmp2;
-    wpts[3] = tmp3;
-    wpts[4] = tmp4;
-    wpts[5] = tmp5;
-    wpts[6] = tmp6;
-    wpts[7] = tmp7;
-    wpts[8] = tmp8;
-}
-
-// Window of width 4
-
-fn window(qpt: &ECp, ppt: &mut ECp) {
-    double_2(ppt);
-    double_2(ppt);
-    double_2(ppt);
-    double_3(ppt);
-    add_2(qpt, ppt);
-}
-
-// Constant time table look-up - borrowed from ed25519 
-
-fn fe_cmov(f: &mut Fq51, g: &Fq51, ib: i8) {
-    let b = -ib as i64;
-    let Fq51(fv) = f;
-    let Fq51(gv) = g;
-    fv[0] ^= (fv[0] ^ gv[0]) & b;
-    fv[1] ^= (fv[1] ^ gv[1]) & b;
-    fv[2] ^= (fv[2] ^ gv[2]) & b;
-    fv[3] ^= (fv[3] ^ gv[3]) & b;
-    fv[4] ^= (fv[4] ^ gv[4]) & b;
-}
-
-fn cmov(w: &mut ECp, u: &ECp, b: i8) {
-  fe_cmov(&mut w.x, &u.x, b);
-  fe_cmov(&mut w.y, &u.y, b);
-  fe_cmov(&mut w.z, &u.z, b);
-  fe_cmov(&mut w.t, &u.t, b);
-}
-
-// return 1 if b==c, no branching
-// must ensure that b, c >= 0
-// does not work for negative arguments
-// used only for windowing purposes with input values 0..8
-fn equal(b: u8, c: u8) -> i8 {
-    let x = ((b ^ c) as i8) - 1; // if (b^c) = 0, x now -1
-    ((x >> 7) & 1)
-}
-
-fn select(tpt: &mut ECp, wpts: &[ECp], b: i8) {
-    // incoming window values range from [-8..+8)
-    let m = b >> 7;
-    let babs = ((b ^ m) - m) as u8;
-
-    cmov(tpt, &wpts[0], equal(babs,0));  // conditional move
-    cmov(tpt, &wpts[1], equal(babs,1));
-    cmov(tpt, &wpts[2], equal(babs,2));
-    cmov(tpt, &wpts[3], equal(babs,3));
-    cmov(tpt, &wpts[4], equal(babs,4));
-    cmov(tpt, &wpts[5], equal(babs,5));
-    cmov(tpt, &wpts[6], equal(babs,6));
-    cmov(tpt, &wpts[7], equal(babs,7));
-    cmov(tpt, &wpts[8], equal(babs,8)); 
-    
-    let mtpt = -(*tpt);
-    cmov(tpt, &mtpt, m & 1);
-}
-
-// Point Multiplication - exponent is 251 bits
-
-// multiply incoming projective point, 
-// return result in projective coords
-fn mul_to_proj(w: &WinVec, ppt: &mut ECp) {
-    let mut wpts = [PT_INF; NPREP];
-    precomp(ppt, &mut wpts);
-
-    let WinVec(wv) = w;
-    let ix = w.0[PANES-1] as usize;
-    *ppt = wpts[ix];
-    let mut jx = PANES-1;
-    let mut qpt = PT_INF;
-    for _ in 0..(PANES-1) {
-        jx -= 1;
-        select(&mut qpt, &wpts, wv[jx]);
-        window(&qpt, ppt);
-    }
-}
-
-// multiply incoming projective point,
-// return result in affine coords
-fn ecp_mul(w: &WinVec, ppt: &mut ECp) {
-    mul_to_proj(w, ppt);
-    norm(ppt);
-}
-
-// convert incoming LEV32 (byte vector) into a little endian vector
-// of bipolar window values [-8..8)
-fn cwin4(q: &LEV32, w: &mut WinVec) {
-    // convert incoming N to bipolar 4-bit window vector - no branching
-    let mut cy = 0;
-    let mut cvbip = | v_in | {
-        let mut v = cy + (v_in as i8);
-        cy = v >> 3;
-        cy |= cy >> 1;
-        cy &= 1;
-        v -= cy << 4;
-        v
-    };
-    
-    for ix in 0..32 {
-        let byt = q.0[ix];
-        let v = cvbip(byt & 15);
-        let jx = 2*ix;
-        w.0[jx] = v;
-        let v = cvbip(byt >> 4);
-        w.0[jx+1] = v;
-    }
-}
-
-// point additon of two projective points
-fn add_proj(qpt: &ECp, ppt: &ECp, zpt: &mut ECp) {
-    // Add Q to P, both in projective (X,Y,Z) coordinates. We don't use T here.
-    let a = qpt.z * ppt.z;
-    let b = a.sqr();
-    let c = qpt.x * ppt.x;
-    let d = qpt.y * ppt.y;
-    let e = CURVE_D * c * d;
-    let f = b - e;
-    let g = b + e;
-    let x3 = qpt.x + qpt.y;
-    let y3 = ppt.x + ppt.y;
-    let z3 = x3 * y3;
-    let y3 = z3 - c;
-    let x3 = y3 - d;
-    let y3 = f * x3;
-    let x3 = a * y3;
-    let y3 = d - c;
-    let z3 = g * y3;
-    let y3 = a * z3;
-    let z3 = f * g; // c = 1
-    zpt.x = x3;
-    zpt.y = y3;
-    zpt.z = z3;
-    zpt.t = Fq51::zero();
-}
-
-// -------------------------------------------------------------------------------
-// Binary/frames conversions
-
-// convert a frame Fq51 from 51-bit representation into
-// into a collection of 64-bit cells
-fn elt_to_bin(x: &Fq51, y: &mut U256) {
-    let mut xx = *x;
-    scr(&mut xx);
-    let mut s  = xx.0[0] as u128;
-    s += (xx.0[1] as u128) << 51;
-    y.0[0] = s as u64;
-    s >>= 64;
-    s += (xx.0[2] as u128) << (2*51-64);
-    y.0[1] = s as u64;
-    s >>= 64;
-    s += (xx.0[3] as u128) << (3*51-128);
-    y.0[2] = s as u64;
-    s >>= 64;
-    s += (xx.0[4] as u128) << (4*51-192);
-    y.0[3] = s as u64;
-}
-
-// convert consecutive (little-endian) 64-bit cells
-// into 51-bit representation
-fn bin_to_elt(y: &U256, x: &mut Fq51) {
-    {
-        let mut s = y.0[0] as u128;
-        x.0[0] = (s as i64) & BOT_51_BITS;
-        s >>= 51;
-        s += (y.0[1] as u128) << (64-51);
-        x.0[1] = (s as i64) & BOT_51_BITS;
-        s >>= 51;
-        s += (y.0[2] as u128) << (128-2*51);
-        x.0[2] = (s as i64) & BOT_51_BITS;
-        s >>= 51;
-        s += (y.0[3] as u128) << (192-3*51);
-        x.0[3] = (s as i64) & BOT_51_BITS;
-        s >>= 51;
-        x.0[4] = s as i64;
-    }
-    scr(x);
-}
-
-// -----------------------------------------------------------------------
-// Input string conversions
-
-// collect a vector of 8-bit values from a hex string.
-// the vector has little-endian order
-fn str_to_bin8(s: &str, x: &mut [u8]) {
-    let nx = x.len();
-    let mut bf = 0;
-    let mut bw = 0;
-    let mut val: u8 = 0;
-    for c in s.chars().rev() {
-        match c.to_digit(16) {
-            Some(d) => {
-                val |= (d as u8) << bf;
-                bf += 4;
-                if bf == 8 {
-                    if bw < nx {
-                        x[bw] = val;
-                    }
-                    bf = 0;
-                    bw += 1;
-                    val = 0;
-                }
-            },
-            None => panic!("Invalid hex digit")
-        }
-    }
-    if bf > 0 && bw < nx {
-        x[bw] = val;
-    }
-}
-
-// collect a vector of 64-bit cells from a hex string
-// the vector has little-endian order
-fn str_to_bin64(s: &str, x: &mut [u64]) {
-    let nx = x.len();
-    let mut bf = 0;
-    let mut bw = 0;
-    let mut val: u64 = 0;
-    for c in s.chars().rev() {
-        match c.to_digit(16) {
-            Some(d) => {
-                val |= (d as u64) << bf;
-                bf += 4;
-                if bf == 64 {
-                    if bw < nx {
-                        x[bw] = val;
-                    }
-                    bf = 0;
-                    bw += 1;
-                    val = 0;
-                }
-            },
-            None => panic!("Invalid hex digit")
-        }
-    }
-    if bf > 0 && bw < nx {
-        x[bw] = val;
-    }
-}
-
-fn str_to_Fq(s: &str) -> Fq {
-    let mut bin : [u64;4] = [0;4];
-    str_to_bin64(s, &mut bin);
-    let mut ans = U256(bin);
-    while ans >= Q {
-        sub_noborrow(&mut ans.0, &Q.0);
-    }
-    Fq::from(FqUnscaled(ans))
-}
-
-fn str_to_Fr(s: &str) -> Fr {
-    let mut bin : [u64;4] = [0;4];
-    str_to_bin64(s, &mut bin);
-    let mut ans = U256(bin);
-    while ans >= R {
-        sub_noborrow(&mut ans.0, &R.0);
-    }
-    Fr::from(FrUnscaled(ans))
-}
-
-// convert bignm string to Fq51
-fn str_to_elt(s: &str, e: &mut Fq51) {
-    let mut bin: [u64;4] = [0;4];
-    str_to_bin64(s, &mut bin);
-    bin_to_elt(&U256(bin), e);
-}
-
-// convert bignum string to 4-bit LE window vector
-fn str_to_winvec(s: &str, w: &mut WinVec) {
-    let mut qv: [u8;32] = [0;32];
-    str_to_bin8(s, &mut qv);
-    println!("multiplier: {}", LEV32(qv));
-    cwin4(&LEV32(qv), w);
-}
+mod cpt;
+use self::cpt::*;
 
 // -------------------------------------------------------
 #[cfg(test)]
@@ -2214,115 +154,6 @@ mod tests {
 }
 
 // ------------------------------------------------------------------------------------------
-// Client API - compressed points and simple fields
-
-#[derive(Copy, Clone, Eq, PartialEq)]
-pub struct Pt([u8;32]);
-
-impl Pt {
-    pub fn bits(self) -> [u8;32] {
-        self.0
-    }
-
-    fn nbr_str(&self) -> String {
-        let Pt(qv) = self;
-        let v = unsafe { mem::transmute::<[u8;32], [u64;4]>(*qv) };
-        basic_nbr_str(&v)
-    }
-
-    fn basic_from_ECp(pt : &ECp) -> Pt {
-        // only for affine pts
-        let ptx = FqUnscaled::from(pt.x);
-        let LEV32(mut x) = LEV32::from(U256::from(ptx));
-        if pt.y.is_odd() {
-            x[31] |= 0x80;
-        }
-        Pt(x)
-    }
-}
-
-impl fmt::Display for Pt {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "Pt({})", self.nbr_str())
-    }
-}
-
-impl From<ECp> for Pt {
-    fn from(pt : ECp) -> Pt {
-        let mut afpt = pt;
-        if !afpt.is_affine() {
-            norm(&mut afpt);
-        }
-        Pt::basic_from_ECp(&afpt)
-    }
-}
-
-pub fn pt_on_curve(pt : Pt) -> Option<ECp> {
-    let mut x = pt.bits();
-    let sgn = (x[31] & 0x80) != 0;
-    x[31] &= 0x7f;
-    if x == [0u8;32] { None } // can't be pt at infinity
-    else {
-        let xq = Fq51::from(FqUnscaled(U256::from(LEV32(x))));
-        match solve_ypt(xq) {
-            Some(ysqrt) => {
-                let yq = if ysqrt.is_odd() == sgn { ysqrt } else { -ysqrt };
-                let newpt = ECp::from_xy(xq, yq);
-                // check for small subgroup attack
-                if (CURVE_H * newpt).is_inf() { None } else { Some(newpt) }
-            },
-            None => None
-        }
-    }
-}
-
-fn solve_ypt(xq : Fq51) -> Option<Fq51> {
-    let yyq = ((1 + xq) * (1 - xq)) / (1 - xq.sqr() * CURVE_D);
-    gsqrt(yyq)
-}
-
-impl From<Pt> for ECp {
-    fn from(pt : Pt) -> ECp {
-        // NOTE: This will panic if invalid point
-        pt_on_curve(pt).unwrap()
-    }
-}
-
-impl From<Hash> for ECp {
-    fn from(h : Hash) -> ECp {
-        let mut bits = h.bits();
-        let sgn = bits[31] & 0x80 != 0;
-        bits[31] &= 0x7f;
-        let mut x = U256::from(LEV32(bits));
-        while x >= Q {
-            div2(&mut x.0);
-        }
-        let mut xq = Fq51::from(FqUnscaled(x));
-        let hpt : ECp;
-        loop {
-            if xq != Fq51::zero() { 
-                // can't be pt at inf
-                match solve_ypt(xq) {
-                    Some(ysqrt) => {
-                        // xq must be on curve
-                        let yq = if sgn == ysqrt.is_odd() { ysqrt } else { -ysqrt };
-                        // avoid small subgroup
-                        let pt = CURVE_H * ECp::from_xy(xq, yq);
-                        if !pt.is_inf() {
-                            hpt = pt;
-                            break;
-                        }
-                    },
-                    None => {}
-                }
-            }
-            xq = xq.sqr() + 1;
-        }
-        hpt
-    }
-}
-
-// ------------------------------------------------------------------------------------------
 
 pub fn curve1174_tests() {
     let smul = "1FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF77965C4DFD307348944D45FD166C970"; // *ed-r* - 1
@@ -2342,32 +173,6 @@ pub fn curve1174_tests() {
     let pt1 = ECp::new(gx, gy);
     let pt2 = pt1 + pt1;
     println!("ptsum {}", pt2);
-
-    // --------------------------------------------
-    // multiplier in 4-bit window form
-    let mut w = WINVEC_INIT;
-    str_to_winvec(smul, &mut w);
-
-    let mut gen_x = Fq51::zero();
-    let mut gen_y = Fq51::zero();
-    str_to_elt(sx, &mut gen_x); // *ed-gen*
-    str_to_elt(sy, &mut gen_y);
-
-    println!("The Generator Point");
-    println!("gen_x: {}", gen_x);
-    println!("gen_y: {}", gen_y);
-
-    let mut pt1 = PT_INF;
-    for _ in 0..100 {
-        init(&gen_x, &gen_y, &mut pt1);
-        ecp_mul(&w, &mut pt1);
-    }
-   
-    println!("Result as Bignums");
-    println!("pt: {}", pt1);
-    
-    println!("Result as Fq51s");
-    println!("pt: {:#?}", pt1);
 
     let tmp = Fr::from(2);
     let tmp2 = 1 / tmp;
@@ -2397,6 +202,15 @@ pub fn curve1174_tests() {
     let out = hasher.result();
     println!("{:x}", out);
     */
+    let mut gen_x = Fq51::zero();
+    let mut gen_y = Fq51::zero();
+    str_to_elt(sx, &mut gen_x); // *ed-gen*
+    str_to_elt(sy, &mut gen_y);
+
+    println!("The Generator Point");
+    println!("gen_x: {}", gen_x);
+    println!("gen_y: {}", gen_y);
+
     println!("x+y: {}", gen_x + gen_y);
     /* */
     let ept = ECp::from(hash(b"Testing12")); // produces an odd Y
